Adds server-client implement that can send a single file to localhost.
This is the start of a server-client setup for QBART. The end product has
a master PYNQ as a client, and 0+ slave PYNQs running as python servers.

For now, this is a simple start, which can send a single file from client to
server, and confirm that the file is received.

What it does, specifically:
* Client has a hardcoded IP, port and image path that it uses.
* The server also is hardcoded with an IP, port, and image path.
* Everything is set up so that generalization to several images, and several servers,
should be a breeze.
* Start the server, then start the client. The client will connect to the server, and transfer
the test image.