Adds client-server setup to MVP Notebook
This implementation is much closer to the final version. Here we pass pickled image lists
and pickled QNN from client to server. Server processes, and then responds with the category
number classification for each image.

Lots of bugs fixed. Mostly with declaring the right sizes for object sending.
Also, the default python pickle does not work well with large objects, so we use
cPickle instead, which solves a very strange bug during development where the pickle
with the normal pickle implement wouldn't work as before.

Now this must be tested on one to several PYNQs, threading, and ordered result return
must be implemented.