Adds server status message functionality.
Now each server will send status messages back to master PYNQ to tell when
they have finished processing an image. Master PYNQ can then pass on these messages
to the microcontroller on the PCB after cffi integration. It will be up to the application
how one wishes to display the status updates. It might quickly become more relevant to show
percentages rather than the exact image that it is working on. We'll have to test to find out what works best!